requirements1.txt
Django==1.7.4
Whoosh==2.5.7
django-cms==3.0.10
django-codemirror-widget==0.4.0
django-debug-toolbar==0.8.5
django-djconfig==0.1.8
django-haystack==2.3.1
django-hvad==0.4.0
django-infinite-scroll-pagination==0.1.3
django-mandrill==0.1.0
django-mptt==0.6.1
django-reversion==1.8.5
django-rosetta==0.7.4
django-spirit==0.1.3
django-wysiwyg-redactor==0.4.5
djangocms-admin-style==0.2.5
mandrill==1.0.57
python-slugify==0.1.0
pytils==0.3
pytz==2014.10
sorl-thumbnail==12.2
